trustlens/backend/requirements.txt
opencv-contrib-python
numpy
Pillow
mediapipe
pytesseract
pyzbar
scikit-image
uvicorn
fastapi
python-multipart
exifread
requests